rename Clio.c and Clio.h to clio.c and clio.h
