+++ libfreedo/arm.c
@@ -31,7 +31,7 @@
 
 #include "arm.h"
 #include "Madam.h"
-#include "Clio.h"
+#include "clio.h"
 #include "DiagPort.h"
 #include "sport.h"
