+++ libfreedo/xbus.c
@@ -28,7 +28,7 @@
   *  Felix Lazarev
 */
 
-#include "Clio.h"
+#include "clio.h"
 #include "xbus.h"
 
 #include <stdint.h>
